requirements.txt
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.7
comm==0.1.2
contourpy==1.2.0
cycler==0.10.0
Cython==3.0.5
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.6
django-rest-framework==0.1.0
djangorestframework==3.14.0
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
fastapi==0.108.0
fastjsonschema==2.16.2
filelock==3.12.4
fonttools==4.47.0
fqdn==1.5.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.42
grad-cam==1.5.0
h11==0.14.0
huggingface-hub==0.20.3
idna==2.10
imbalanced-learn==0.12.0
imblearn==0.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab==3.5.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
kiwisolver==1.4.5
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==2.0.4
mock-open==1.4.0
mpmath==1.3.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.3
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==22.0
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.2
Pillow==10.1.0
platformdirs==3.10.0
polars==0.20.3
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pycairo==1.26.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pygame==2.5.2
Pygments==2.13.0
pyparsing==3.1.1
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.4
python-magic==0.4.27
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
regex==2023.12.25
requests==2.28.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.11.4
seaborn==0.13.1
Send2Trash==1.8.0
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.32.0.post1
sympy==1.12
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.15.1
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tornado==6.2
tqdm==4.66.1
traitlets==5.7.1
transformers==4.37.2
ttach==0.0.3
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.1.40
uri-template==1.2.0
urllib3==1.26.13
uvicorn==0.25.0
virtualenv==20.24.5
wandb==0.16.3
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
